Fix bug with `null` pull request number on branch push
When running the action on branch pushes, `jq .number ${GITHUB_EVENT_PATH}` is actually returning `null` instead of an empty value. This evaluates to the string `"null"` in the bash variable, which in turns makes the condition of the if statement truthy.

If `INPUT_NETLIFY_SITE` is specified, the action fails: it's trying to evaluate a pull request preview for number PR `null`, which return a 404.